fix compiler warnings 1
```
src/misc.c:519:24: warning: implicit conversion loses integer precision: '__darwin_time_t' (aka 'long') to 'int' [-Wshorten-64-to-32]
src/session.c:665:52: warning: implicit conversion loses integer precision: 'long' to '__darwin_suseconds_t' (aka 'int') [-Wshorten-64-to-32]
```

+++ src/misc.c
@@ -483,7 +483,7 @@ _libssh2_debug_low(LIBSSH2_SESSION * session, int context, const char *format,
     int len, msglen, buflen = sizeof(buffer);
     va_list vargs;
     struct timeval now;
-    static int firstsec;
+    static long firstsec;
     static const char *const contexts[] = {
         "Unknown",
         "Transport",
